+++ tests/FlowsTest.php
@@ -1,9 +1,9 @@
 <?php
 
-use Wistrix\Onboard\Flows;
-use Wistrix\Onboard\Step;
-use Wistrix\Onboard\Tests\Stubs\Group;
-use Wistrix\Onboard\Tests\Stubs\User;
+use Xoiatix\Onboard\Flows;
+use Xoiatix\Onboard\Step;
+use Xoiatix\Onboard\Tests\Stubs\Group;
+use Xoiatix\Onboard\Tests\Stubs\User;
 
 beforeEach(function () {
     $this->user = new User;