+++ CHANGELOG.md
@@ -2,13 +2,17 @@
 
 All notable changes will be documented in this file.
 
+## 1.2.0 - 2025-06-16
+
+- Namespace updated from `Wistrix` to `Xoiatix`
+
 ## 1.1.1 - 2024-04-28
 
 - Aligning middleware tests step naming and validate callbacks
 
 ## 1.1.0 - 2024-04-27
 
-- Added `Wistrix\Onboard\Facades\Onboard` facade
+- Added `Xoiatix\Onboard\Facades\Onboard` facade
 - Allow steps to be restricted to specific classes
 - Removed ability to register steps via middleware